Move `CurrentPane` to the file named `current_pane.rs`
