+++ src/usecase/fzf_make/current_pane.rs
@@ -1,0 +1,15 @@
+#[derive(Clone, PartialEq, Debug)]
+pub enum CurrentPane {
+    Main,
+    History,
+}
+
+impl CurrentPane {
+    pub fn is_main(&self) -> bool {
+        matches!(self, CurrentPane::Main)
+    }
+
+    pub fn is_history(&self) -> bool {
+        matches!(self, CurrentPane::History)
+    }
+}
