+++ src/usecase/fzf_make/mod.rs
@@ -1,2 +1,3 @@
 pub(super) mod app;
+pub(super) mod current_pane;
 pub(super) mod ui;
